requirements.txt
aiogram==3.26.0
python-dotenv==1.0.0
aiohttp==3.13.3
asyncio==4.0.0
logging==0.4.9.6
yoomoney==2.0.0